+++ src/main/java/com/verdantartifice/primalmagic/common/rituals/UniversalRitualStep.java
@@ -2,6 +2,7 @@ package com.verdantartifice.primalmagic.common.rituals;
 
 import net.minecraft.core.BlockPos;
 import net.minecraft.nbt.CompoundTag;
+import net.minecraft.resources.ResourceLocation;
 
 /**
  * Class identifying a single step in a ritual's process, one targeting a nearby universal prop.
@@ -10,24 +11,31 @@ import net.minecraft.nbt.CompoundTag;
  */
 public class UniversalRitualStep extends AbstractRitualStep {
     protected BlockPos pos;
+    protected ResourceLocation expectedId;
     
     public UniversalRitualStep() {
         super();
-        this.pos = BlockPos.ZERO;
+        this.pos = null;
+        this.expectedId = null;
     }
     
-    public UniversalRitualStep(BlockPos pos) {
+    public UniversalRitualStep(BlockPos pos, ResourceLocation expectedId) {
         super(RitualStepType.UNIVERSAL_PROP);
         this.pos = pos;
+        this.expectedId = expectedId;
     }
 
     @Override
     public boolean isValid() {
-        return super.isValid() && this.pos != null;
+        return super.isValid() && this.pos != null && this.expectedId != null;
     }
     
     public BlockPos getPos() {
         return this.pos;
+    }
+    
+    public ResourceLocation getExpectedId() {
+        return this.expectedId;
     }
 
     @Override
@@ -36,6 +44,7 @@ public class UniversalRitualStep extends AbstractRitualStep {
         retVal.putInt("PosX", this.pos.getX());
         retVal.putInt("PosY", this.pos.getY());
         retVal.putInt("PosZ", this.pos.getZ());
+        retVal.putString("ExpectedId", this.expectedId.toString());
         return retVal;
     }
 
@@ -43,5 +52,6 @@ public class UniversalRitualStep extends AbstractRitualStep {
     public void deserializeNBT(CompoundTag nbt) {
         super.deserializeNBT(nbt);
         this.pos = new BlockPos(nbt.getInt("PosX"), nbt.getInt("PosY"), nbt.getInt("PosZ"));
+        this.expectedId = new ResourceLocation(nbt.getString("ExpectedId"));
     }
 }